backend/pyproject.toml
[project]
name = "recipefy"
version = "0.1.0"
description = "Application to create recipes with AI"
authors = [
    {name = "Mateusz Kocik",email = "matkot11@gmail.com"}
]
requires-python = "==3.14.1"
dependencies = [
    "django==5.2.8",
    "djangorestframework==3.16.1",
    "django-cors-headers==4.3.1",
    "psycopg>=3.1.0",
    "environs==14.5.0",
    "django-redis==6.0.0",
    "drf-spectacular==0.29.0",
]

[project.optional-dependencies]
dev = [
    "ruff==0.14.6",
    "pytest",
    "pytest-django",
]

[dependency-groups]
dev = [
    "pytest==9.0.1",
    "pytest-django==4.11.1",
    "pytest-cov==6.0.0"
]
